+++ bayesnet/classifiers/KDBLd.cc
@@ -7,7 +7,25 @@
 #include "KDBLd.h"
 
 namespace bayesnet {
-    KDBLd::KDBLd(int k) : KDB(k), Proposal(dataset, features, className) {}
+    KDBLd::KDBLd(int k) : KDB(k), Proposal(dataset, features, className)
+    {
+        validHyperparameters = validHyperparameters_ld;
+        validHyperparameters.push_back("k");
+        validHyperparameters.push_back("theta");
+    }
+    void KDBLd::setHyperparameters(const nlohmann::json& hyperparameters_)
+    {
+        auto hyperparameters = hyperparameters_;
+        if (hyperparameters.contains("k")) {
+            k = hyperparameters["k"];
+            hyperparameters.erase("k");
+        }
+        if (hyperparameters.contains("theta")) {
+            theta = hyperparameters["theta"];
+            hyperparameters.erase("theta");
+        }
+        Proposal::setHyperparameters(hyperparameters);
+    }
     KDBLd& KDBLd::fit(torch::Tensor& X_, torch::Tensor& y_, const std::vector<std::string>& features_, const std::string& className_, map<std::string, std::vector<int>>& states_, const Smoothing_t smoothing)
     {
         checkInput(X_, y_);