+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2025-06-30
+
+### Internal
+
+- Add docs generation to CMakeLists.txt.
+- Add new hyperparameters to the Ld classifiers:
+  - *ld_algorithm*: algorithm to use for local discretization, with the following options: "MDLP", "BINQ", "BINU".
+  - *ld_proposed_cuts*: number of cut points to return.
+  - *mdlp_min_length*: minimum length of a partition in MDLP algorithm to be evaluated for partition.
+  - *mdlp_max_depth*: maximum level of recursion in MDLP algorithm.
+
 ## [1.1.1] - 2025-05-20
 
 ### Internal